src/api-server/requirements.txt
flask~=3.0.2
joblib~=1.3.2
scikit-learn~=1.4.1.post1
pandas~=2.2.1
scipy~=1.12.0
flask-cors~=4.0.1

tensorflow==2.16.1
tensorflow-hub==0.16.1
nltk~=3.8.1
beautifulsoup4~=4.12.3
soupsieve~=2.5
h11~=0.14.0
jupyter~=1.0.0
attrs~=23.2.0
wheel~=0.38.4
tornado~=6.4
Jinja2~=3.1.3
pyzmq~=26.0.3
colorama~=0.4.6
setuptools~=65.5.1
comm~=0.2.2
traitlets~=5.14.3
fqdn~=1.5.1
gast~=0.5.4
numpy~=1.26.4
h5py~=3.11.0
ipython~=8.24.0
parso~=0.8.4
jedi~=0.19.1
pytz~=2024.1
QtPy~=2.4.1
packaging~=24.0
rich~=13.7.1
pygments~=2.18.0
PyYAML~=6.0.1
sniffio~=1.3.1
arrow~=1.3.0
python-dateutil~=2.9.0.post0
Babel~=2.15.0
MarkupSafe~=2.1.5
werkzeug~=3.0.1
click~=8.1.7
blinker~=1.7.0
itsdangerous~=2.2.0
httpcore~=1.0.5
idna~=3.7
certifi~=2024.2.2
keras~=3.3.3
optree~=0.11.0
requests~=2.31.0
namex~=0.0.8
mdurl~=0.1.2
overrides~=7.7.0
cffi~=1.16.0
six~=1.16.0
threadpoolctl~=3.4.0
platformdirs~=4.2.1
webencodings~=0.5.1
bleach~=6.1.0
tinycss2~=1.3.0
psutil~=5.9.8
debugpy~=1.8.1
decorator~=5.1.1
regex~=2024.5.10
wcwidth~=0.2.13
Markdown~=3.6
urllib3~=2.2.1
anyio~=4.3.0
nbformat~=5.10.4
nbclient~=0.10.0
fastjsonschema~=2.19.1
jsonschema~=4.22.0
jupyterlab~=4.1.8
notebook~=7.1.3
wrapt~=1.16.0
asttokens~=2.4.1
executing~=2.0.1
ipykernel~=6.29.4
nbconvert~=7.16.4
pandocfilters~=1.5.1
defusedxml~=0.7.1
mistune~=3.0.2
pycparser~=2.22
qtconsole~=5.5.2
termcolor~=2.4.0
websocket-client~=1.8.0
terminado~=0.18.1
ipywidgets~=8.1.2
referencing~=0.35.1
webcolors~=1.13
jsonpointer~=2.4
isoduration~=20.11.0
httpx~=0.27.0
Send2Trash~=1.8.3
flatbuffers~=24.3.25
astunparse~=1.6.3
tensorboard~=2.16.2
json5~=0.9.25
gensim~=4.3.2
transformers~=4.40.2
filelock~=3.14.0
tqdm~=4.66.4
fsspec~=2024.3.1
torch~=2.3.0
mpmath~=1.3.0
sympy~=1.12